rename: move the tests
