+++ package.json
@@ -5,8 +5,8 @@
   "main": "src/index.js",
   "scripts": {
     "start:dev": "node --trace-warnings --watch --trace-sigint src/index.js",
-    "test:dev": "NODE_ENV=test node --watch --test",
-    "test:prod": "NODE_ENV=test node --test",
+    "test:dev": "NODE_ENV=test node --watch --test tests/**/*.test.js",
+    "test:prod": "NODE_ENV=test node --test --test tests/**/*.test.js",
     "start:prod": "node src/index.js"
   },
   "type": "module",
