+++ templates/sendgrid/error.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Job finalized with success!</title>
+  <title>Job finalized with error!</title>
   <style>
     body {
       font-family: Arial, sans-serif;
